refactor: switch from tabs to pages
