+++ layouts/hub-page.tsx
@@ -27,7 +27,7 @@ const HubPageLayout = ({ children }: { children: React.ReactNode }) => {
   return (
     <div className="flex flex-col min-h-screen">
       <TopNav />
-      <div className="page-container flex grow bg-light-slate-3 flex-col items-center pt-20 md:pt-14">
+      <div className="flex flex-col items-center pt-20 page-container grow bg-light-slate-3 md:pt-14">
         <div className="info-container container w-full min-h-[100px]">
           <Header>
             {insight ? (
